Add interfaces.ts files to fix the interfaces file structure
