+++ frontend/app/src/components/App/interfaces.ts
@@ -1,0 +1,9 @@
+import { SingleRoute } from '../../interfaces';
+
+export interface AppProps {
+  routes: SingleRoute[];
+}
+
+export interface StateProps {
+  isLoading: boolean;
+}
