+++ frontend/app/src/components/RouteWithSubRoutes/interfaces.ts
@@ -1,0 +1,10 @@
+import { SingleRoute } from '../../interfaces';
+import React from 'react';
+
+export interface RouteWithSubRoutesProps {
+  path: string;
+  exact: boolean;
+  routes?: SingleRoute[];
+  component: React.ElementRef<any>;
+  key?: string;
+}
