+++ frontend/app/src/interfaces.ts
@@ -1,0 +1,6 @@
+export interface SingleRoute {
+  path: string;
+  exact: boolean;
+  component: any;
+  routes?: SingleRoute[];
+}
