+++ frontend/app/src/scenes/Home/interfaces.ts
@@ -1,0 +1,11 @@
+import { SingleRoute } from '../../interfaces';
+
+export interface HomeProps {
+  routes: SingleRoute[];
+}
+
+export interface University {
+  id: number;
+  name: string;
+  imageUrl: string;
+}
