+++ frontend/app/src/scenes/Home/scenes/Ambassadors/scenes/AddAmbassador/components/SearchUniversity/interfaces.ts
@@ -1,0 +1,12 @@
+import { University } from '../../../../../../interfaces';
+
+export interface SearchUniversityProps {
+  isVisible: boolean;
+  onSelect: (university: University) => void;
+  onCancel: () => void;
+}
+
+export interface SearchUniversityStateProps {
+  universities: University[];
+  isLoading: boolean;
+}
