+++ frontend/app/src/scenes/Home/scenes/Ambassadors/scenes/AddAmbassador/interfaces.ts
@@ -1,0 +1,11 @@
+import { University } from '../../../../interfaces';
+
+export interface AddAmbassadorStateProps {
+  isLoading: boolean;
+  university: University | null;
+  isSearchUniversityModalVisible: boolean;
+}
+
+export interface AddAmbassadorPayload {
+  id: number;
+}
