+++ frontend/app/src/scenes/Home/scenes/Universities/interfaces.ts
@@ -1,0 +1,6 @@
+import { University } from '../../interfaces';
+
+export interface UniversitiesStateProps {
+  universities: University[];
+  isLoading: boolean;
+}
