+++ frontend/app/src/scenes/Home/scenes/Universities/scenes/AddUniversity/interfaces.ts
@@ -1,0 +1,3 @@
+export interface AddUniversityStateProps {
+  isLoading: boolean;
+}
