+++ frontend/app/src/scenes/Login/interfaces.ts
@@ -1,0 +1,7 @@
+export interface LoginPayload {
+  username: string;
+}
+
+export interface LoginStateProps {
+  isIncorrect: boolean;
+}
